+++ src/translations/pwdmanager-qt_fr.ts
@@ -1,0 +1,473 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE TS>
+<TS version="2.1" language="fr_FR">
+<context>
+    <name>CreateUser</name>
+    <message>
+        <location filename="../createuser.cpp" line="5"/>
+        <source>New Credentials Wizard</source>
+        <translation>Ajouter une  entrée</translation>
+    </message>
+    <message>
+        <location filename="../createuser.cpp" line="10"/>
+        <source>Username: </source>
+        <translation>Nom d&apos;utilisateur: </translation>
+    </message>
+    <message>
+        <location filename="../createuser.cpp" line="16"/>
+        <source>Used for: </source>
+        <translation>Utilisation: </translation>
+    </message>
+    <message>
+        <location filename="../createuser.cpp" line="22"/>
+        <source>Enter your password below, or generate a new one: </source>
+        <translation>Entrez votre mot de passe ci-dessous (ou générez-en un nouveau): </translation>
+    </message>
+    <message>
+        <location filename="../createuser.cpp" line="28"/>
+        <source>Length:</source>
+        <translation>Longueur:</translation>
+    </message>
+    <message>
+        <location filename="../createuser.cpp" line="36"/>
+        <source>Uppercase characters</source>
+        <translation>Majuscules</translation>
+    </message>
+    <message>
+        <location filename="../createuser.cpp" line="38"/>
+        <source>Numeric characters</source>
+        <translation>Chiffres</translation>
+    </message>
+    <message>
+        <location filename="../createuser.cpp" line="40"/>
+        <source>Special characters</source>
+        <translation>Caractères spéciaux</translation>
+    </message>
+    <message>
+        <location filename="../createuser.cpp" line="42"/>
+        <source>Generate</source>
+        <translation>Générer</translation>
+    </message>
+    <message>
+        <location filename="../createuser.cpp" line="51"/>
+        <source>Validate</source>
+        <translation>Valider</translation>
+    </message>
+    <message>
+        <location filename="../createuser.cpp" line="53"/>
+        <source>Cancel</source>
+        <translation>Annuler</translation>
+    </message>
+</context>
+<context>
+    <name>MainWindow</name>
+    <message>
+        <location filename="../database.cpp" line="25"/>
+        <source>Invalid file format</source>
+        <translation>Format de fichier invalide</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="47"/>
+        <source>Invalid header signature</source>
+        <translation>Signature de l&apos;en-tête invalide</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="76"/>
+        <source>Read </source>
+        <translation>Lu </translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="76"/>
+        <source> lines</source>
+        <translation> entrées</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="147"/>
+        <source>Wrote 1 entry.</source>
+        <translation>Écrit 1 entrée.</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="149"/>
+        <source>Wrote </source>
+        <translation>Écrit </translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="149"/>
+        <location filename="../mainwindow.cpp" line="193"/>
+        <location filename="../mainwindow.cpp" line="229"/>
+        <location filename="../mainwindow.cpp" line="298"/>
+        <source> entries.</source>
+        <translation> entrées.</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="164"/>
+        <location filename="../database.cpp" line="382"/>
+        <location filename="../database.cpp" line="458"/>
+        <source>Warning</source>
+        <translation>Attention</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="223"/>
+        <source>Enter a new username:</source>
+        <translation>Entrez un nouveau nom d&apos;utilisateur:</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="226"/>
+        <source>Enter what these credentials will be used for:</source>
+        <translation>Entrez le contexte d&apos;utilisation de ces informations:</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="382"/>
+        <source>This will replace all entries in the current database with those of the imported database.
+This operation is not recoverable. Do you wish to proceed ?</source>
+        <translation>Ceci va remplacer toutes les entrées dans la base de données actuelle avec celles de la base de données importée.
+Cette opération est irréversible. Êtes vous sûr de vouloir continuer ?</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="398"/>
+        <source>Invalid file</source>
+        <translation>Fichier invalide</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="398"/>
+        <source>This file does not appear to be a valid .crdb file !</source>
+        <translation>Ce fichier ne semble pas être un fichier .crdb valide !</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="438"/>
+        <source>Exported database to
+</source>
+        <translation>Exporté la base de données vers
+</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="269"/>
+        <source>
+to clipboard</source>
+        <translation>
+vers le presse-papiers</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="164"/>
+        <source>This will delete the entire row from the database.
+This operation is not recoverable. Do you wish to proceed ?</source>
+        <translation>Cette ligne sera effacée de la base de données.
+Cette opération est irréversible. Êtes-vous sûr de vouloir continuer ?</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="223"/>
+        <location filename="../database.cpp" line="226"/>
+        <location filename="../mainwindow.cpp" line="137"/>
+        <source>Edit</source>
+        <translation>Éditer</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="269"/>
+        <source>Copied password from row </source>
+        <translation>Copié le mot de passe de la ligne </translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="17"/>
+        <source>&amp;File</source>
+        <translation>&amp;Fichier</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="18"/>
+        <source>Import</source>
+        <translation>Importer</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="22"/>
+        <source>Export</source>
+        <translation>Exporter</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="26"/>
+        <source>Restore from backup</source>
+        <translation>Restorer depuis la copie de sauvegarde</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="30"/>
+        <source>Convert old file</source>
+        <translation>Convertir un ancien format</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="34"/>
+        <source>Refresh View</source>
+        <translation>Actualiser</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="39"/>
+        <source>Quit</source>
+        <translation>Quitter</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="43"/>
+        <source>&amp;Edit</source>
+        <translation>&amp;Éditer</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="44"/>
+        <source>New row</source>
+        <translation>Nouvelle entrée</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="48"/>
+        <source>Delete row</source>
+        <translation>Effacer</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="52"/>
+        <source>Edit cell content</source>
+        <translation>Éditer cellule</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="56"/>
+        <source>Copy password</source>
+        <translation>Copier</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="60"/>
+        <source>Move row up</source>
+        <translation>Déplacer ligne vers le haut</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="65"/>
+        <source>Move row down</source>
+        <translation>Déplacer ligne vers le bas</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="71"/>
+        <location filename="../misc.cpp" line="5"/>
+        <source>Help</source>
+        <translation>Aide</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="76"/>
+        <location filename="../misc.cpp" line="27"/>
+        <source>About</source>
+        <translation>À propos</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="85"/>
+        <source>USERNAME;USED FOR;PASSWORD</source>
+        <translation>UTILISATEUR;UTILISE POUR;MOT DE PASSE</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="99"/>
+        <source>Search:</source>
+        <translation>Rechercher:</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="108"/>
+        <source>E&amp;xact match</source>
+        <translation>Correspondance e&amp;xacte</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="113"/>
+        <source>Clear</source>
+        <translation>Effacer</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="129"/>
+        <source>New Entry</source>
+        <translation>Nouveau</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="133"/>
+        <source>Delete</source>
+        <translation>Effacer ligne</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="141"/>
+        <source>Copy</source>
+        <translation>Copier</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="189"/>
+        <location filename="../mainwindow.cpp" line="225"/>
+        <source>Retrieved 0 entries.</source>
+        <translation>Lu 0 entrée.</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="191"/>
+        <location filename="../mainwindow.cpp" line="227"/>
+        <source>Retrieved 1 entry.</source>
+        <translation>Lu 1 entrée.</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="193"/>
+        <location filename="../mainwindow.cpp" line="229"/>
+        <location filename="../mainwindow.cpp" line="298"/>
+        <source>Retrieved </source>
+        <translation>Lu </translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="221"/>
+        <source>Hidden table view.
+Press Ctrl+H to unhide</source>
+        <translation>Vue masquée.
+Appuyez sur Ctrl+H pour démasquer</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="330"/>
+        <location filename="../mainwindow.cpp" line="358"/>
+        <source>No matches found.</source>
+        <translation>Aucune correspondance trouvée.</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="332"/>
+        <location filename="../mainwindow.cpp" line="360"/>
+        <source>Found 1 match.</source>
+        <translation>Trouvé 1 correspondance.</translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="334"/>
+        <location filename="../mainwindow.cpp" line="362"/>
+        <source>Found </source>
+        <translation>Trouvé </translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="334"/>
+        <location filename="../mainwindow.cpp" line="362"/>
+        <source> matches.</source>
+        <translation> correspondances.</translation>
+    </message>
+    <message>
+        <location filename="../misc.cpp" line="5"/>
+        <source>Add a new row with New Entry.
+Double click on a cell, or highlight it then press the Edit button to edit its contents.
+Highlight a cell and press Delete to erase the entire row.
+Copy copies the highlighted row&apos;s password field into your OS&apos; clipboard.
+Move rows around with Ctrl+Up or Ctrl+Down.
+
+The search field looks for a case-insensitive match of the string entered. It is, however, accent-sensitive (e.g. it will not match &quot;Umlaut&quot; and &quot;Ümlaut&quot;).
+Press Ctrl+H to hide/unhide the table (if someone is spying on you !)
+
+In case things go horribly wrong during a transaction (e.g. power loss), you can find a backup of the database in the application data (%AppData%\Roaming\passwordmanager on Windows, $HOME/.passwordmanager on Linux) as a .bak file. Import it in the manager, and you should get your passwords back like they were before the transaction.</source>
+        <translation>Ajoutez une nouvelle entrée avec le bouton Nouveau.
+Double cliquez sur une cellule, où cliquez puis pressez le bouton Éditer pour éditer le contenu.
+Cliquez sur une cellule puis appuyez sur Effacer pour effacer toute la ligne.
+Copier copie le mot de passe de la ligne séléctionnée dans le presse-papiers.
+Déplacez les lignes vers le haut ou vers le bas à l&apos;aide du raccourci Ctrl+Haut ou Ctrl+Bas.
+
+La fonction de recherche n&apos;est pas sensible à la casse. Cependant, elle est sensible aux accents (par exemple, elle ne trouvera pas de correspondance entre &quot;Umlaut&quot; et &quot;Ümlaut&quot;.)
+Appuyez sur Ctrl+H pour masquer/démasquer la base de données (par exemple si quelqu&apos;un regarde par dessus votre épaule !)
+
+Dans le cas  où une transaction serait interrompue (par exemple dû à une panne de courant), une copie de sauvegarde de la base de données est disponible dans le dossier de l&apos;application (%AppData%\Roaming\passwordmanager sous Windows, $HOME/.passwordmanager sous Linux) en tant que fichier.bak. Importez ce fichier dans l&apos;application et vous devriez retrouver vos mots de passe.</translation>
+    </message>
+    <message>
+        <location filename="../misc.cpp" line="21"/>
+        <source>pwdmanager-qt version </source>
+        <translation>pwdmanager-qt version </translation>
+    </message>
+    <message>
+        <location filename="../misc.cpp" line="21"/>
+        <source> by Lou VINCENT aka Bad64.
+</source>
+        <translation>par Lou VINCENT aka Bad64.
+</translation>
+    </message>
+    <message>
+        <location filename="../misc.cpp" line="22"/>
+        <source>Made with Qt </source>
+        <translation>Crée avec Qt </translation>
+    </message>
+    <message>
+        <location filename="../misc.cpp" line="22"/>
+        <source>, with GCC version </source>
+        <translation>, avec GCC version </translation>
+    </message>
+    <message>
+        <location filename="../misc.cpp" line="22"/>
+        <source>, based on pwdmanager (also by Bad64).
+
+</source>
+        <translation>, basé sur pwdmanager (également par Bad64).
+
+</translation>
+    </message>
+    <message>
+        <location filename="../misc.cpp" line="23"/>
+        <source>This software is provided as-is, free of charge, for personal use.
+</source>
+        <translation>Ce programme est fourni tel quel, gratuitement, pour un usage personnel.
+</translation>
+    </message>
+    <message>
+        <location filename="../misc.cpp" line="24"/>
+        <source>Bad64 shall not be held responsible for passwords stolen by coworkers looking above one&apos;s shoulder, </source>
+        <translation>Bad64 ne saurait être tenu responsables du vol de mots de passe par des collègues regardant par dessus votre épaule, </translation>
+    </message>
+    <message>
+        <location filename="../misc.cpp" line="25"/>
+        <source>loss of important credentials, or broken coffee machines.</source>
+        <translation>la perte d&apos;authentifiants importants, ou les pannes survennant sur la machine à café.</translation>
+    </message>
+    <message>
+        <location filename="../database.cpp" line="458"/>
+        <source>The software will now attempt to convert this file to the new format.
+This operation may fail if the file is malformed. Do you wish to proceed ?</source>
+        <translation>Le programme va maintenant tenter de convertir ce fichier vers le nouveau format.
+Cette opération peut échouer si le fichier est malformé. Désirez-vous continuer ?</translation>
+    </message>
+    <message>
+        <location filename="../askuserpassword.cpp" line="8"/>
+        <source>Enter password</source>
+        <translation>Mot de passe</translation>
+    </message>
+    <message>
+        <location filename="../askuserpassword.cpp" line="8"/>
+        <source>Please enter your Windows account password:</source>
+        <translation>Veuillez saisir votre mot de passe de votre compte Windows:</translation>
+    </message>
+</context>
+<context>
+    <name>MiniGenerateBox</name>
+    <message>
+        <location filename="../minigeneratebox.cpp" line="5"/>
+        <source>Edit password</source>
+        <translation>Editer le mot de passe</translation>
+    </message>
+    <message>
+        <location filename="../minigeneratebox.cpp" line="13"/>
+        <source>Enter a new password (or generate one):</source>
+        <translation>Entrez un nouveau mot de passe (ou générez-en un):</translation>
+    </message>
+    <message>
+        <location filename="../minigeneratebox.cpp" line="20"/>
+        <source>Length:</source>
+        <translation>Longueur:</translation>
+    </message>
+    <message>
+        <location filename="../minigeneratebox.cpp" line="28"/>
+        <source>Uppercase characters</source>
+        <translation>Majuscules</translation>
+    </message>
+    <message>
+        <location filename="../minigeneratebox.cpp" line="30"/>
+        <source>Numeric characters</source>
+        <translation>Chiffres</translation>
+    </message>
+    <message>
+        <location filename="../minigeneratebox.cpp" line="32"/>
+        <source>Special characters</source>
+        <translation>Caractères spéciaux</translation>
+    </message>
+    <message>
+        <location filename="../minigeneratebox.cpp" line="34"/>
+        <source>Generate</source>
+        <translation>Générer</translation>
+    </message>
+    <message>
+        <location filename="../minigeneratebox.cpp" line="43"/>
+        <source>Validate</source>
+        <translation>Valider</translation>
+    </message>
+    <message>
+        <location filename="../minigeneratebox.cpp" line="45"/>
+        <source>Cancel</source>
+        <translation>Annuler</translation>
+    </message>
+</context>
+</TS>